requirements.txt
# Qlib 核心框架
pyqlib>=0.9.0

# 数据源
akshare>=1.11.0
efinance>=0.5.0

# 数据处理
pandas>=1.5.0
numpy>=1.23.0

# 可视化
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0

# 数据库（可选）
sqlalchemy>=2.0.0

# 工具
tqdm>=4.64.0
pyyaml>=6.0
loguru>=0.6.0

# Jupyter支持
jupyter>=1.0.0
ipython>=8.0.0

# 技术指标
ta-lib>=0.4.0  # 可选，需要单独安装底层库
pandas-ta>=0.3.0  # 替代方案
